feat(AlertBlock.java): Add abstraction to the alert block classes
